[#9][#10] M vs C plugins and appropriate install directories
This commit takes care of three issues:

- Plug-ins written in M need to install in `$ydb_dist/plugin/o[/utf8]`.
This is now in the variable YOTTADB_M_PLUGIN_DIR.
- Plug-ins written in C need to install in `$ydb_dist/plugin`. This is
now in the variable YOTTADB_C_PLUGIN_DIR.
- The plugin install directory for M plug-ins was delegated to the
end user. This is not optimal. Now YOTTADB_M_PLUGIN_DIR points to the
correct install directory for the plugin based on wheather UTF-8 mode was
requested or not.

Also, `ci/needs_copyright.sh` needs to return `1` not `0` for files
in this repo.