[#12] `-embed_source` must be the default
Due to limitations on how YottaDB works (see
https://gitlab.com/YottaDB/DB/YDB/-/issues/909 for full discussion),
compiled libraries are not guarenteed to have access to the original
sources (which may be deleted/modified later, or only kept in a temp
directory that is later removed). We need to embed the source in the
objects until we have a working solution at the YottaDB level.