[#5] No M directory for M mode code + install_ydb_sources() + README update
We found out that in Octo, we have a convention where the M mode code is
in the current binary directory (not in the M folder), whereas the utf8
mode code is in the utf8 folder. So we are removing the M folder to
avoid having to change the scripts that Octo has. This also matches more
or less how YottaDB works, where M mode code is in the root folder, and
utf8 code is in the utf8 folder.

I also took the opportunity to add the function `install_ydb_sources()`
which does what it says. It's technically useless (as the source code is
not used at runtime), but we have a convention right now of installing the
useless source code.

These changes will necessitate a small change in the YDBPosix repo
because the tests need to run after build time, and the location of the
M mode .so file will be different.

README was also updated with an advanced section for using a local copy
of this repository and debugging information.