[#13] Remove unused M_UTF8_MODE variable
This was missed from #5. It's no longer used.

There is currently a bug there where setting the old variable
M_UTF8_MODE causes the base of the installation to be
${plugin_dir}/utf8, as a result, the M mode code gets installed in the
utf8 directory, and the UTF-8 code gets installed in the utf8/utf8
directory. As a result, if M_UTF8_MODE is used, and since UTF-8 mode is
installed last, the utf8/ folder object ends up being the M mode object.
Better to just remove the variable now.