[#14] Fix regression causing YDB-E-NONUTF8LOCALE
Turns out, there is a regression in the fix done for #14 in !24. If you
do cmake for a second or subsequent time, the value of LC_ALL is lost,
potentially resulting in this error: `%YDB-E-NONUTF8LOCALE, Locale has
character encoding (ANSI_X3.4-1968) which is not compatible with UTF-8
character set.` This error does not typically show up on our work
machines, as the default character set is already set to be UTF-8, so
losing LC_ALL effectively results in LC_ALL being a UTF-8 character set;
but on Docker pipelines, the default locale is C, and we see this issue.

The solution is to make LC_ALL a cached but internal variable (not
changable in CMake GUIs).

Pipeline code in `ci/test_ydb_container.sh` has been enhanced to show
that this problem exists.